requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0

# Database
mysqlclient>=2.2,<3.0

# Redis
redis>=5.0,<6.0

# Production server
gunicorn>=21.0,<23.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<6.0

# Code quality
flake8>=7.0,<8.0
black>=24.0,<25.0
isort>=5.13,<6.0
